Fixes some docker issues
Changes alpine to use a tag instead of a sha hash. The image associated with
a tag can change, which can be undesirable. In this case, it's probably fine,
but it's something to remember when working with docker image tags.

Changed some indentation

Added InitialWorkDirRequirement to samtools

samtools tries to write a FASTA index to the same directory the FASTA file is in.
Inside a container this might be read-only. So by adding the InitialWorkDirRequirement
we make sure the FASTA file is mounted in a writeable directory.